+++ UIComponents.Abstractions/Interfaces/FileExplorer/IUICFileExplorerService.cs
@@ -23,5 +23,12 @@ namespace UIComponents.Abstractions.Interfaces.FileExplorer
         Task RenameFileOrDirectoryAsync(RelativePathModel pathModel, string newName);
 
         Task DownloadFilesAndDirectories(List<RelativePathModel> pathModels, Stream outputStream);
+
+        /// <summary>
+        /// When trying to upload a file, this method can provide a override of the file uploader. return Task.FromResult(null) to use the default file uploader.
+        /// </summary>
+        /// <param name="pathModel"></param>
+        /// <returns></returns>
+        Task<IUIComponent?> GetUploadFileComponent(RelativePathModel pathModel);
     }
 }