+++ UIComponents.Generators/Services/UICFileExplorerService.cs
@@ -845,7 +845,12 @@ public class UICFileExplorerService : IUICFileExplorerService
         return fileInfo.Extension != string.Empty;
     }
 
-   
+    public Task<IUIComponent?> GetUploadFileComponent(RelativePathModel pathModel)
+    {
+        return Task.FromResult<IUIComponent?>(null);
+    }
+
+
     #endregion