+++ UIComponents.Models/Models/Questions/UICQuestionText.cs
@@ -12,6 +12,21 @@ public class UICQuestionText : UICQuestionBase<string>
     {
     }
 
+    public bool Multiline { get; set; }
+
+    /// <summary>
+    /// Validate the text result before submitting.
+    /// return a string if there is a validation error
+    /// </summary>
+    /// <remarks>
+    /// Available args: value
+    /// </remarks>
+    public IUICAction TextValidation { get; set; } = new UICCustom();
+
+    public bool ValidateRequired { get; set; } = true;
+    public int? ValidateMinLength { get; set; } 
+    public int? ValidateMaxLength { get; set; }
+
     public override string MapResponse(string response) => response;
 
     public static UICQuestionText Create(Translatable title, Translatable message, IUICQuestionService questionService)